refactor: move multer config in a dedicated file in utils
