+++ src/controllers/productController.ts
@@ -1,28 +1,7 @@
 import { Request, Response } from "express";
 import ProductModel, { IProductDocument } from "../models/productModel";
 import mongoose from "mongoose";
-import multer from "multer";
 import path from "path";
-import fs from "fs";
-
-export const directoryToStoreImages = "uploads/products/images";
-
-// Create the directory if it doesn't exist
-if (!fs.existsSync(directoryToStoreImages)) {
-  fs.mkdirSync(directoryToStoreImages, { recursive: true });
-}
-
-// Multer configuration
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, directoryToStoreImages); // Specify the directory to store uploaded files
-  },
-  filename: function (req, file, cb) {
-    cb(null, Date.now() + "-" + file.originalname); // Rename the file to avoid conflicts
-  },
-});
-
-export const upload = multer({ storage: storage });
 
 /**
  * Controller for handling product-related operations.
